+++ .github/workflows/main.yml
@@ -5,9 +5,10 @@ on: [push]
 jobs:
   test-node:
     runs-on: ubuntu-latest
+    timeout-minutes: 10
     strategy:
       matrix:
-        node-version: [12.x, 14.x]
+        node-version: [14.x, 16.x, 18.x]
     steps:
     - uses: actions/checkout@v2
     - name: Use Node.js ${{ matrix.node-version }}
@@ -17,13 +18,27 @@ jobs:
     - run: npm install
     - name: Run test with Node.js ${{ matrix.node-version }}
       run: npm run test-node
-      env:
-        CI: true
-  lint:
+  test-karma:
     runs-on: ubuntu-latest
+    timeout-minutes: 10
     strategy:
       matrix:
-        node-version: [14.x]
+        node-version: [16.x]
+    steps:
+    - uses: actions/checkout@v2
+    - name: Use Node.js ${{ matrix.node-version }}
+      uses: actions/setup-node@v1
+      with:
+        node-version: ${{ matrix.node-version }}
+    - run: npm install
+    - name: Run karma tests
+      run: npm run test-karma
+  lint:
+    runs-on: ubuntu-latest
+    timeout-minutes: 10
+    strategy:
+      matrix:
+        node-version: [16.x]
     steps:
     - uses: actions/checkout@v2
     - name: Use Node.js ${{ matrix.node-version }}
@@ -34,11 +49,12 @@ jobs:
     - name: Run eslint
       run: npm run lint
   coverage:
-    needs: [test-node]
+    needs: [test-node, test-karma]
     runs-on: ubuntu-latest
+    timeout-minutes: 10
     strategy:
       matrix:
-        node-version: [14.x]
+        node-version: [16.x]
     steps:
     - uses: actions/checkout@v2
     - name: Use Node.js ${{ matrix.node-version }}
@@ -48,10 +64,8 @@ jobs:
     - run: npm install
     - name: Generate coverage report
       run: npm run coverage-ci
-      env:
-        CI: true
     - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v1
+      uses: codecov/codecov-action@v2
       with:
-        file: ./coverage.lcov
+        file: ./coverage/lcov.info
         fail_ci_if_error: true
