--- lib/main.js
@@ -1,20 +1,0 @@
-/*!
- * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
- */
-
-import {DidKeyDriver} from './DidKeyDriver.js';
-
-/**
- * Helper method to match the `.driver()` API of other `did-io` plugins.
- *
- * @param {object} options - Options hashmap.
- * @param {object} [options.verificationSuite=Ed25519VerificationKey2020] -
- *   Key suite for the signature verification key suite to use.
- *
- * @returns {DidKeyDriver} Returns an instance of a did:key resolver driver.
- */
-function driver({verificationSuite} = {}) {
-  return new DidKeyDriver({verificationSuite});
-}
-
-export {driver, DidKeyDriver};
