+++ package.json
@@ -2,7 +2,7 @@
   "name": "@digitalbazaar/did-method-key",
   "version": "2.0.1-0",
   "description": "A did:key method resolver.",
-  "homepage": "http://github.com/digitalbazaar/did-method-key",
+  "homepage": "https://github.com/digitalbazaar/did-method-key",
   "author": {
     "name": "Digital Bazaar, Inc.",
     "email": "support@digitalbazaar.com",
@@ -17,40 +17,38 @@
     "email": "support@digitalbazaar.com"
   },
   "license": "BSD-3-Clause",
-  "main": "lib/index.js",
-  "module": "lib/main.js",
+  "type": "module",
+  "exports": "./lib/index.js",
   "dependencies": {
-    "@digitalbazaar/did-io": "^1.0.0",
-    "@digitalbazaar/ed25519-verification-key-2020": "^3.0.0",
-    "@digitalbazaar/x25519-key-agreement-key-2019": "^5.1.1",
-    "@digitalbazaar/x25519-key-agreement-key-2020": "^2.0.0",
-    "esm": "^3.2.25"
+    "@digitalbazaar/did-io": "^2.0.0",
+    "@digitalbazaar/ed25519-verification-key-2020": "^4.0.0",
+    "@digitalbazaar/x25519-key-agreement-key-2019": "^6.0.0",
+    "@digitalbazaar/x25519-key-agreement-key-2020": "^3.0.0"
   },
   "devDependencies": {
-    "@digitalbazaar/ed25519-verification-key-2018": "^3.1.1",
-    "chai": "^4.3.4",
+    "@digitalbazaar/ed25519-verification-key-2018": "^4.0.0",
+    "c8": "^7.11.2",
+    "chai": "^4.3.6",
     "cross-env": "^7.0.3",
-    "eslint": "^7.22.0",
-    "eslint-config-digitalbazaar": "^2.6.1",
-    "eslint-plugin-jsdoc": "^32.3.0",
-    "karma": "^6.3.1",
-    "karma-babel-preprocessor": "^8.0.1",
+    "eslint": "^8.14.0",
+    "eslint-config-digitalbazaar": "^3.0.0",
+    "eslint-plugin-jsdoc": "^39.2.9",
+    "eslint-plugin-unicorn": "^42.0.0",
+    "karma": "^6.3.19",
+    "karma-babel-preprocessor": "^8.0.2",
     "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^3.1.0",
+    "karma-chrome-launcher": "^3.1.1",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
     "karma-sourcemap-loader": "^0.3.8",
     "karma-webpack": "^5.0.0",
-    "mocha": "^8.3.2",
+    "mocha": "^9.2.2",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^15.1.0",
-    "webpack": "^5.28.0",
-    "webpack-bundle-analyzer": "^4.4.0",
-    "webpack-cli": "^4.5.0"
+    "webpack": "^5.72.0"
   },
-  "nyc": {
+  "c8": {
     "exclude": [
-      "tests"
+      "test"
     ],
     "reporter": [
       "html",
@@ -63,7 +61,7 @@
     "util": false
   },
   "engines": {
-    "node": ">=12"
+    "node": ">=14"
   },
   "keywords": [
     "Decentralized",
@@ -72,11 +70,11 @@
   ],
   "scripts": {
     "test": "npm run lint && npm run test-node && npm run test-karma",
-    "test-node": "cross-env NODE_ENV=test mocha -r esm --preserve-symlinks -t 10000 test/*.spec.js",
-    "test-karma": "karma start karma.conf.js",
-    "coverage": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text-summary npm run test-node",
-    "coverage-ci": "cross-env NODE_ENV=test nyc --reporter=text-lcov npm run test-node > coverage.lcov",
-    "coverage-report": "nyc report",
+    "test-node": "cross-env NODE_ENV=test mocha --preserve-symlinks -t 10000 test/*.spec.js",
+    "test-karma": "karma start karma.conf.cjs",
+    "coverage": "cross-env NODE_ENV=test c8 --reporter=lcov --reporter=text-summary npm run test-node",
+    "coverage-ci": "cross-env NODE_ENV=test c8 --reporter=lcovonly --reporter=text npm run test-node",
+    "coverage-report": "c8 report",
     "lint": "eslint ."
   }
 }
